requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SocketIO==5.3.6
WTForms==3.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
python-socketio==5.8.0
eventlet==0.33.3
bcrypt==4.0.1
python-dotenv==1.0.0
email_validator==2.0.0
beautifulsoup4==4.12.2
